Simple Shell written in C
Written in C. Accepts arguments. README explains capabilities and
possible shortcomings. Built with Wesley Minner. Makefile allows you to
make a dist, test it and clean the directory. make time allows you to
compare the times and resources this program takes to complete the same
commands as the shell and execline.